Async awesomeness
Ok so loading those json files really needs to be protected in some way.

I didn't implement promises but what I've done is pretty close. Although
at the end of it I descovered that each .js file has a .filePath property
on it which could be used to easily create relative paths. Making the
hardcoded root information kind of pointless. Will need to revisit all
of this logic to see if I can make it even simpler/smarter by using that
.filePath prop in the _this_ for each .js file. Good news is the current
implementation works really well and is kind of cool :)